// typechain-types/IPredictionMarketFactory.ts
export type MarketCreationDataStruct = {
  marketType: BigNumberish;
  creator: AddressLike;
  resolver: AddressLike;
  feeRecipient: AddressLike;
  platformFeeBps: BigNumberish;
  question: string;
  description: string;
  category: BigNumberish;
  subcategory: BigNumberish;
  endTime: BigNumberish;
};